Parameter Make.2-T

type t

The type of terms.

type location

The type of locations.

val atom : ?⁠loc:location -> int -> t

Make an atom from an non-zero integer. Positive integers denotes variables, and negative integers denote the negation of the variable corresponding to their absolute value.