Module Make.Parser

The Parser module for the language.

Main interface

exception Error

Exception raised by the parser when it encounters an error.

val file : (Stdlib.Lexing.lexbuf -> token) -> Stdlib.Lexing.lexbuf -> statement list

A function that parses an entire file, i.e until the end-of-file token, and return the list of parsed statements.

val input : (Stdlib.Lexing.lexbuf -> token) -> Stdlib.Lexing.lexbuf -> statement option

A function to parse a single statement. Reutnrs None if it encounters the end-of-file token. Used for incremental parsing.

Menhir incremental interface

module MenhirInterpreter : sig ... end

module Incremental : sig ... end