Module S.Lexer

The Lexer module for the language.

type token

The type of token produced by the lexer.

exception Error

The exception raised by the lexer when it cannot produce a token.

val descr : token -> Tok.descr

Asspcoate a description to each token, to help with error messages.

val token : (Stdlib.Lexing.lexbuf -> unit) -> Stdlib.Lexing.lexbuf -> token

The function producing token from a lexbuf.