Module type Parse.S

Main interface

type token

The type of token consumed by the parser.

type statement

The type of top-level declarations returned by the parser.

exception Error of int

Exception raised by the parser when it encounters an error.

val file : (Stdlib.Lexing.lexbuf -> token) -> Stdlib.Lexing.lexbuf -> statement list

A function that parses an entire file, i.e until the end-of-file token, and return the list of parsed statements.

val input : (Stdlib.Lexing.lexbuf -> token) -> Stdlib.Lexing.lexbuf -> statement option

A function to parse a single statement. Returns None if it encounters the end-of-file token. Used for incremental parsing.