Module Dolmen_intf.Stmt

Interfaces for statements. This module defines interfaces for statements, i.e top-level declarations in files.

module type Response = sig ... end

Signature used by the Response class, which parses prover responses, i.e. SAT/UNSAT answers (with option certificates).

module type Logic = sig ... end

Signature used by the Logic class, which parses languages such as tptp, smtlib, etc... Statements of dirrent languages currently have a lot less in common than terms, so this interface looks a lot more like a patchwork of different logical framework directives than it should.