Module Dolmen_intf.Term

Signature for Parsing Logic languages

module type Logic = sig ... end

Signature for Typechecked terms

module type Tff = sig ... end

module type Ae_Base = sig ... end

Minimum required to type ae's tff

module type Ae_Arith = sig ... end

Minimum required to type ae's arith

module type Tptp_Base = sig ... end

Minimum required to type tptp's tff

module type Tptp_Arith_Common = sig ... end

Common signature for tptp arithmetics

module type Tptp_Arith = sig ... end

Signature required by terms for typing tptp arithmetic.

module type Smtlib_Base = sig ... end

Minimum required to type smtlib's core theory.

module type Smtlib_Arith_Common = sig ... end

Common signature for first-order arithmetic

module type Smtlib_Int = sig ... end

Signature required by terms for typing smtlib int arithmetic.

module type Smtlib_Real = sig ... end

Signature required by terms for typing smtlib real arithmetic.

module type Smtlib_Real_Int = sig ... end

Signature required by terms for typing smtlib real_int arithmetic.

module type Smtlib_Array = sig ... end

module type Smtlib_Bitv = sig ... end

module type Smtlib_Float_Bitv = sig ... end

Bitvector part of the smtlib float requirements

module type Smtlib_Float_Real = sig ... end

Real part of the smtlib float requirements

module type Smtlib_Float_Float = sig ... end

Float part of the smtlib float requirements

module type Smtlib_Float = sig ... end

module type Smtlib_String_RegLan = sig ... end

module type Smtlib_String_String = sig ... end

module type Smtlib_String = sig ... end