Module type Term.Smtlib_Base

Minimum required to type smtlib's core theory.

type t

The type of terms.

type cstr

The type of ADT constructor

val eqs : t list -> t

Create a chain of equalities.

val cstr_tester : cstr -> t -> t

Given a constructor c and a term t, returns a terms that evaluates to true iff t has c as head constructor.