Module type Ty.Smtlib_Array

Signature required by types for typing smtlib arrays

type t

The type of types

val array : t -> t -> t

The type of functionnal arrays from one type to another.

type view#row

Partial views for types. These are used in the Array theory to enforce some restrictions logics impose on the types of arrays that cna occur.

and view = private [>

| `Int

| `Real

| `Bitv of int

| `Array of t * t

]

Partial views for types. These are used in the Array theory to enforce some restrictions logics impose on the types of arrays that cna occur.

val view : t -> view

Partial view of a type.