Module Dolmen_intf.View

First-order View

module Sexpr : sig ... end

This explicits a view of untyped terms as first-order terms/formulas

First-order View

module TFF : sig ... end

This explicits a view of terms as polymorphic first-order types and terms. In strict first-order, the distinction between types and terms is clear and useful enough that the view distinguishes the two.