Module View.TFF

This explicits a view of terms as polymorphic first-order types and terms. In strict first-order, the distinction between types and terms is clear and useful enough that the view distinguishes the two.

module Sig : sig ... end

module Ty : sig ... end

module TypeDef : sig ... end

module Term : sig ... end

module Vty = Ty

module VT = Term

module type S = sig ... end

The signature for a module that provides a first-order view of terms and types.