Module TFF.Ty

type _ view =

| Var : 'ty_var -> < ty_var : 'ty_var.. > view

| App : 'ty_cst * 'ty list -> < ty_cst : 'ty_cst ; builtin : 'blt ; ty : 'ty.. > view

View of types in first-order. In first-order, types, which include basically variables and application of type constructors to types, are differentiated from type signatures, which include function types.