Module type Dolmen_intf__Term.Smtlib_Float

type t

The type of terms

type ty

The type of types.

val ty : t -> ty

Type of a term.

module Real : Smtlib_Float_Real with type t := t

Sub-module used for namespacing the real part of the theory requirements

module Bitv : Smtlib_Float_Bitv with type t := t

Sub-module used for namespacing the bitvector part of the theory requirements

module Float : Smtlib_Float_Float with type t := t

Sub-module used for namespacing the floating number part of the theory requirements