Module type Dolmen_intf__Ty.Smtlib_Real_Int

Signature required for types for typing smtlib real_int arithmetic.

include Smtlib_Int

type t

The type of types

val int : t

The type for integer expressions.

include Smtlib_Real with type t := t

type t

The type of types

val real : t

The type for integer expressions.

type view#row

Partial view for types. These are used by the Reals_Ints theory to perform type-based dispatch, and automatic conversion of Ints to Reals when specified by the specification.

and view = private [>

| `Int

| `Real

]

Partial view for types. These are used by the Reals_Ints theory to perform type-based dispatch, and automatic conversion of Ints to Reals when specified by the specification.

val view : t -> view

Partial view of a type.