Module Builtin.Adt

type _ t =

| Tester : {

adt : 'ty_cst;

case : int;

cstr : 'term_cst;

} -> < ty_cst : 'ty_cst ; term_cst : 'term_cst.. > t
(*
Tester { adt; cstr; case; } is the tester of the case-th constructor of type adt which should be cstr.
*)

| Constructor : {

adt : 'ty_cst;

case : int;

} -> < ty_cst : 'ty_cst.. > t
(*
Constructor { adt; case} is the case-th constructor of the algebraic datatype defined by adt.
*)

| Destructor : {

adt : 'ty_cst;

case : int;

cstr : 'term_cst;

field : int;

} -> < ty_cst : 'ty_cst ; term_cst : 'term_cst.. > t
(*
Destructor { adt; cstr; case; field; } is the destructor returning the field-th argument of the case-th constructor of type adt which should be cstr.
*)