Module Builtin.Arith

type _ t =

| Int
(*
Int: ttype the type for signed integers of arbitrary precision.
*)

| Integer of string
(*
Integer s: Int: integer literal. The string s should be the decimal representation of an integer with arbitrary precision (hence the use of strings rather than the limited precision int).
*)

| Rat
(*
Rat: ttype the type for signed rationals.
*)

| Rational of string
(*
Rational s: Rational: rational literal. The string s should be the decimal representation of a rational (see the various languages spec for more information).
*)

| Real
(*
Real: ttype the type for signed reals.
*)

| Decimal of string
(*
Decimal s: Real: real literals. The string s should be a floating point representation of a real. Not however that reals here means the mathematical abstract notion of real numbers, including irrational, non-algebric numbers, and is thus not restricted to floating point numbers, although these are the only literals supported.

Note that, in spite of the name, the literals may not be expressed in decimal notation. For instance, Decimal "0x1.0p1" is a valid representation for the real number 2.

Real literals can be parsed using ZArith's Q.of_string.
*)

| Lt of [ `Int | `Rat | `Real ]
(*
Lt: {a=(Int|Rational|Real)} a -> a -> Prop.T: strict comparison (less than) on numbers (whether integers, rationals, or reals).
*)

| Leq of [ `Int | `Rat | `Real ]
(*
Leq:{a=(Int|Rational|Real)} a -> a -> Prop.T: large comparison (less or equal than) on numbers (whether integers, rationals, or reals).
*)

| Gt of [ `Int | `Rat | `Real ]
(*
Gt:{a=(Int|Rational|Real)} a -> a -> Prop.T: strict comparison (greater than) on numbers (whether integers, rationals, or reals).
*)

| Geq of [ `Int | `Rat | `Real ]
(*
Geq:{a=(Int|Rational|Real)} a -> a -> Prop.T: large comparison (greater or equal than) on numbers (whether integers, rationals, or reals).
*)

| Minus of [ `Int | `Rat | `Real ]
(*
Minus:{a=(Int|Rational|Real)} a -> a: arithmetic unary negation/minus on numbers (whether integers, rationals, or reals).
*)

| Add of [ `Int | `Rat | `Real ]
(*
Add:{a=(Int|Rational|Real)} a -> a -> a: arithmetic addition on numbers (whether integers, rationals, or reals).
*)

| Sub of [ `Int | `Rat | `Real ]
(*
Sub:{a=(Int|Rational|Real)} a -> a -> a: arithmetic substraction on numbers (whether integers, rationals, or reals).
*)

| Mul of [ `Int | `Rat | `Real ]
(*
Mul:{a=(Int|Rational|Real)} a -> a -> a: arithmetic multiplication on numbers (whether integers, rationals, or reals).
*)

| Pow of [ `Int | `Real ]
(*
Pow:{a=(Int|Real)} a -> a -> a: arithmetic exponentiation on numbers (whether integers or reals).
*)

| Div of [ `Rat | `Real ]
(*
Div:{a=(Rational|Real)} a -> a -> a: arithmetic exact division on numbers (rationals, or reals, but **not** integers).
*)

| Div_e of [ `Int | `Rat | `Real ]
(*
Div_e:{a=(Int|Rational|Real)} a -> a -> a: arithmetic integer euclidian quotient (whether integers, rationals, or reals). If D is positive then Div_e (N,D) is the floor (in the type of N and D) of the real division N/D, and if D is negative then Div_e (N,D) is the ceiling of N/D.
*)

| Modulo_e of [ `Int | `Rat | `Real ]
(*
Modulo_e:{a=(Int|Rational|Real)} a -> a -> a: arithmetic integer euclidian remainder (whether integers, rationals, or reals). It is defined by the following equation: Div_e (N, D) * D + Modulo(N, D) = N.
*)

| Div_t of [ `Int | `Rat | `Real ]
(*
Div_t:{a=(Int|Rational|Real)} a -> a -> a: arithmetic integer truncated quotient (whether integers, rationals, or reals). Div_t (N,D) is the truncation of the real division N/D.
*)

| Modulo_t of [ `Int | `Rat | `Real ]
(*
Modulo_t:{a=(Int|Rational|Real)} a -> a -> a: arithmetic integer truncated remainder (whether integers, rationals, or reals). It is defined by the following equation: Div_t (N, D) * D + Modulo_t(N, D) = N.
*)

| Div_f of [ `Int | `Rat | `Real ]
(*
Div_f:{a=(Int|Rational|Real)} a -> a -> a: arithmetic integer floor quotient (whether integers, rationals, or reals). Div_t (N,D) is the floor of the real division N/D.
*)

| Modulo_f of [ `Int | `Rat | `Real ]
(*
Modulo_f:{a=(Int|Rational|Real)} a -> a -> a: arithmetic integer floor remainder (whether integers, rationals, or reals). It is defined by the following equation: Div_f (N, D) * D + Modulo_f(N, D) = N.
*)

| Abs
(*
Abs: Int -> Int: absolute value on integers.
*)

| Divisible
(*
Divisible: Int -> Int -> Prop.T: divisibility predicate on integers. Smtlib restricts applications of this predicate to have a litteral integer for the divisor/second argument.
*)

| Is_int of [ `Int | `Rat | `Real ]
(*
Is_int:{a=(Int|Rational|Real)} a -> Prop.T: integer predicate for numbers: is the given number an integer.
*)

| Is_rat of [ `Int | `Rat | `Real ]
(*
Is_rat:{a=(Int|Rational|Real)} a -> Prop.T: rational predicate for numbers: is the given number an rational.
*)

| Floor of [ `Int | `Rat | `Real ]
(*
Floor:{a=(Int|Rational|Real)} a -> a: floor function on numbers, defined in tptp as the largest integer not greater than the argument.
*)

| Floor_to_int of [ `Rat | `Real ]
(*
Floor_to_int:{a=(Rational|Real)} a -> Int: floor and conversion to integers in a single function. Should return the greatest integer i such that the rational or real intepretation of i is less than, or equal to, the argument.
*)

| Ceiling of [ `Int | `Rat | `Real ]
(*
Ceiling:{a=(Int|Rational|Real)} a -> a: ceiling function on numbers, defined in tptp as the smallest integer not less than the argument.
*)

| Truncate of [ `Int | `Rat | `Real ]
(*
Truncate:{a=(Int|Rational|Real)} a -> a: ceiling function on numbers, defined in tptp as the nearest integer value with magnitude not greater than the absolute value of the argument.
*)

| Round of [ `Int | `Rat | `Real ]
(*
Round:{a=(Int|Rational|Real)} a -> a: rounding function on numbers, defined in tptp as the nearest intger to the argument; when the argument is halfway between two integers, the nearest even integer to the argument.
*)