Module Builtin.Array

type _ t =

| T
(*
Array.T: ttype -> ttype -> ttype: the type constructor for polymorphic functional arrays. An (src, dst) Array is an array from expressions of type src to expressions of type dst. Typically, such arrays are immutables.
*)

| Const
(*
Store: 'a 'b. 'b -> ('a, 'b) Array.T: returns a constant array, which maps any value of type 'a to the given base value.
*)

| Store
(*
Store: 'a 'b. ('a, 'b) Array.T -> 'a -> 'b -> ('a, 'b) Array.T: store operation on arrays. Returns a new array with the key bound to the given value (shadowing the previous value associated to the key).
*)

| Select
(*
Select: 'a 'b. ('a, 'b) Array.T -> 'a -> 'b: select operation on arrays. Returns the value associated to the given key. Typically, functional arrays are complete, i.e. all keys are mapped to a value.
*)