Module Builtin.Bitv

type _ t =

| T of {

n : int;

}
(*
Bitv.T{n}: ttype: type constructor for bitvectors of length n. Ensures that n > 0.
*)

| Binary_lit of string
(*
Lit s: Bitv.T: bitvector litteral. The string s should be a binary representation of bitvectors using characters '0', and '1' (lsb last)
*)

| To_int of {

n : int;

signed : bool;

}
(*
To_int{n; signed; }: Bitv.T{n} -> Int: conversion from bitvectors to signed integers.
*)

| Of_int of {

n : int;

}
(*
Of_int{n}: Int -> Bitv.T{n}: conversion fromm (signed) integers to bitvectors.
*)

| Concat of {

n : int;

m : int;

}
(*
Concat{n;m}: Bitv.T{n} -> Bitv.T{m} -> Bitv.T{n+m}: concatenation operator on bitvectors.
*)

| Extract of {

n : int;

i : int;

j : int;

}
(*
Extract{n; i; j}: Bitv.T{n} -> Bitv.T{i - j + 1}: bitvector extraction, from index j up to i (both included). Ensures that 0 <= j <= i < n.
*)

| Repeat of {

n : int;

k : int;

}
(*
Repeat{n;k}: Bitv.T{n} -> Bitv.T{n*k}: bitvector repeatition.
*)

| Zero_extend of {

n : int;

k : int;

}
(*
Zero_extend{n;k}: Bitv.T{n} -> Bitv.T{n + k}: zero extension for bitvectors (produces a representation of the same unsigned integer).
*)

| Sign_extend of {

n : int;

k : int;

}
(*
Sign_extend{n;k}: Bitv.T{n} -> Bitv.T{n + k}: sign extension for bitvectors ((produces a representation of the same signed integer).
*)

| Rotate_right of {

n : int;

i : int;

}
(*
Rotate_right{n;i}: Bitv.T{n} -> Bitv.T{n}: logical rotate right for bitvectors by i.
*)

| Rotate_left of {

n : int;

i : int;

}
(*
Rotate_left{n;i}: Bitv.T{n} -> Bitv.T{n}: logical rotate left for bitvectors by i.
*)

| Not of {

n : int;

}
(*
Not{n}: Bitv.T{n} -> Bitv.T{n}: bitwise negation for bitvectors.
*)

| And of {

n : int;

}
(*
And{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: bitwise conjunction for bitvectors.
*)

| Or of {

n : int;

}
(*
Or{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: bitwise disjunction for bitvectors.
*)

| Nand of {

n : int;

}
(*
Nand{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: bitwise negated conjunction for bitvectors. Nand s t abbreviates Not (And s t)).
*)

| Nor of {

n : int;

}
(*
Nor{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: bitwise negated disjunction for bitvectors. Nor s t abbreviates Not (Or s t)).
*)

| Xor of {

n : int;

}
(*
Xor{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: bitwise exclusive disjunction for bitvectors. Xor s t abbreviates Or (And s (Not t)) (And (Not s) t) .
*)

| Xnor of {

n : int;

}
(*
Xnor{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: bitwise negated exclusive disjunction for bitvectors. Xnor s t abbreviates Or (And s t) (And (Not s) (Not t)).
*)

| Comp of {

n : int;

}
(*
Comp{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{1}: Returns the constant bitvector "1" is all bits are equal, and the bitvector "0" if not.
*)

| Neg of {

n : int;

}
(*
Neg{n}: Bitv.T{n} -> Bitv.T{n}: 2's complement unary minus.
*)

| Add of {

n : int;

}
(*
Add{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: addition modulo 2^n.
*)

| Sub of {

n : int;

}
(*
Sub{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: 2's complement subtraction modulo 2^n.
*)

| Mul of {

n : int;

}
(*
Mul{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: multiplication modulo 2^n.
*)

| Udiv of {

n : int;

}
(*
Udiv{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: unsigned division, truncating towards 0.
*)

| Urem of {

n : int;

}
(*
Urem{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: unsigned remainder from truncating division.
*)

| Sdiv of {

n : int;

}
(*
Sdiv{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: 2's complement signed division.
*)

| Srem of {

n : int;

}
(*
Srem{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: 2's complement signed remainder (sign follows dividend).
*)

| Smod of {

n : int;

}
(*
Smod{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: 2's complement signed remainder (sign follows divisor).
*)

| Shl of {

n : int;

}
(*
Shl{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: shift left (equivalent to multiplication by 2^x where x is the value of the second argument).
*)

| Lshr of {

n : int;

}
(*
Lshr{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: logical shift right (equivalent to unsigned division by 2^x, where x is the value of the second argument).
*)

| Ashr of {

n : int;

}
(*
Ashr{n}: Bitv.T{n} -> Bitv.T{n} -> Bitv.T{n}: Arithmetic shift right, like logical shift right except that the most significant bits of the result always copy the most significant bit of the first argument.
*)

| Ult of {

n : int;

}
(*
Ult{n}: Bitv.T{n} -> Bitv.T{n} -> Prop.T: binary predicate for unsigned less-than.
*)

| Ule of {

n : int;

}
(*
Ule{n}: Bitv.T{n} -> Bitv.T{n} -> Prop.T: binary predicate for unsigned less than or equal.
*)

| Ugt of {

n : int;

}
(*
Ugt{n}: Bitv.T{n} -> Bitv.T{n} -> Prop.T: binary predicate for unsigned greater-than.
*)

| Uge of {

n : int;

}
(*
Uge{n}: Bitv.T{n} -> Bitv.T{n} -> Prop.T: binary predicate for unsigned greater than or equal.
*)

| Slt of {

n : int;

}
(*
Slt{n}: Bitv.T{n} -> Bitv.T{n} -> Prop.T: binary predicate for signed less-than.
*)

| Sle of {

n : int;

}
(*
Sle{n}: Bitv.T{n} -> Bitv.T{n} -> Prop.T: binary predicate for signed less than or equal.
*)

| Sgt of {

n : int;

}
(*
Sgt{n}: Bitv.T{n} -> Bitv.T{n} -> Prop.T: binary predicate for signed greater-than.
*)

| Sge of {

n : int;

}
(*
Sge{n}: Bitv.T{n} -> Bitv.T{n} -> Prop.T: binary predicate for signed greater than or equal.
*)

| Overflow_neg of {

n : int;

}
(*
Overflow_neg{n} : Bitv.T{n} -> Prop.T: unary overflow predicate for signed unary minus (i.e. returns true if the negation would overflow)
*)

| Overflow_add of {

n : int;

signed : bool;

}
(*
Overflow_add{n;signed} : Bitv.T{n} -> Bitv.T{n} -> Prop.T: binary overflow predicate for signed/unsigned addition (i.e. returns true if the operation would overflow)
*)

| Overflow_sub of {

n : int;

signed : bool;

}
(*
Overflow_sub{n;signed}: Bitv.T{n} -> Bitv.T{n} -> Prop.T binary overflow predicate for signed/unsigned subtraction (i.e. returns [true] if the operation would overflow)
*)

| Overflow_mul of {

n : int;

signed : bool;

}
(*
Overflow_mul{n;signed}: Bitv.T{n} -> Bitv.T{n} -> Prop.T binary overflow predicate for signed/unsigned multiplication (i.e. returns [true] if the operation would overflow)
*)

| Overflow_div of {

n : int;

}
(*
Overflow_div{n}: Bitv.T{n} -> Bitv.T{n} -> Prop.T binary overflow predicate for signed division (i.e. returns [true] if the operation would overflow)
*)