Module Builtin.Map

type _ t =

| T
(*
Map.T: ttype -> ttype -> ttype is the type for encoding higher order functions into first order. An (arg, ret) Map.T is a first-order type for encodings of functions from arg to ret.
*)

| App
(*
App: 'arg 'ret. ('arg, 'ret) Map.T -> 'arg -> 'ret: Application operation for encodings of higher-order functions.
*)