Module Dolmen_std.Escape

Identifier name manipulation

val smap : (int -> char -> string) -> string -> string

umap f provides an equivalent of flat_map on strings. f is given the position of the character in the string (starting from 1), and a character.

val rename : sep:char -> string -> string

A renaming function, which add an increasing number after the given separator.

Identifier escaping

module type Arg = Dolmen_intf.Id.Escape

module Make : functor (Id : Arg) -> sig ... end