Module Dolmen_std.Expr

Type definitions

Common definitions

type hash = private int

type index = private int

type 'a tag = 'a Tag.t

type builtin = ..

type ttype =

| Type
The type of types.

type 'ty id = private {

ty : 'ty;

name : string;

index : index;
unique

builtin : builtin;

mutable tags : Tag.map;

}

The type of identifiers. 'ty is the type for representing the type of the id, 'ty and 't_n are the lengths of arguments as described by the uiltin type.

type ('ttype, 'ty) function_type = {

fun_vars : 'ttype id list;

fun_args : 'ty list;

fun_ret : 'ty;

}

The type for representing function types.

Types

type ty_var = ttype id

Abbreviation for type variables.

and ty_const = (unit, ttype) function_type id

Type symbols have the expected length of their argument encoded.

and ty_descr =

| Var of ty_var
Type variables

| App of ty_const * ty list
Application

Type descriptions.

and ty = {

as_ : ty_var option;

mutable descr : ty_descr;

mutable hash : hash;

mutable tags : Tag.map;

}

Types, which wrap type description with a memoized hash and some tags.

Terms and formulas

type term_var = ty id

Term variables

and term_const = (ttype, ty) function_type id

Term symbols, which encode their expected type and term argument lists lengths.

and pattern = term

patterns are simply terms

and term_descr =

| Var of term_var
Variables

| App of term_const * ty list * term list
Application

| Binder of binder * term
Binders

| Match of term * (pattern * term) list
Pattern matching

Term descriptions.

and binder =

| Exists of ty_var list * term_var list

| Forall of ty_var list * term_var list

| Letin of (term_var * term) list

Binders.

and term = {

ty : ty;

descr : term_descr;

mutable hash : hash;

mutable tags : Tag.map;

}

Term, which wrap term descriptions.

type formula = term

Alias for signature compatibility (with Dolmen_loop.Pipes.Make for instance).

Exceptions

exception Bad_ty_arity of ty_const * ty list

exception Bad_term_arity of term_const * ty list * term list

exception Type_already_defined of ty_const

exception Filter_failed_ty of string * ty * string

exception Filter_failed_term of string * term * string

type builtin += Base | Wildcard

type builtin += Prop | Unit | Univ

type builtin += Coercion

type builtin += True | False | Equal | Distinct | Neg | And | Or | Nand | Nor | Xor | Imply | Equiv

type builtin += Ite

type builtin += Tester of {

cstr : term_const;

} | Constructor of {

adt : ty_const;

case : int;

} | Destructor of {

adt : ty_const;

cstr : term_const;

case : int;

field : int;

}

type builtin += Int | Integer of string | Rat | Rational of string | Real | Decimal of string | Lt | Leq | Gt | Geq | Minus | Add | Sub | Mul | Div | Div_e | Div_t | Div_f | Modulo_e | Modulo_t | Modulo_f | Abs | Divisible | Is_int | Is_rat | Floor | Ceiling | Truncate | Round

type builtin += Array | Store | Select

type builtin += Bitv of int | Bitvec of string | Bitv_concat | Bitv_extract of int * int | Bitv_repeat | Bitv_zero_extend | Bitv_sign_extend | Bitv_rotate_right of int | Bitv_rotate_left of int | Bitv_not | Bitv_and | Bitv_or | Bitv_nand | Bitv_nor | Bitv_xor | Bitv_xnor | Bitv_comp | Bitv_neg | Bitv_add | Bitv_sub | Bitv_mul | Bitv_udiv | Bitv_urem | Bitv_sdiv | Bitv_srem | Bitv_smod | Bitv_shl | Bitv_lshr | Bitv_ashr | Bitv_ult | Bitv_ule | Bitv_ugt | Bitv_uge | Bitv_slt | Bitv_sle | Bitv_sgt | Bitv_sge

type builtin += RoundingMode | RoundNearestTiesToEven | RoundNearestTiesToAway | RoundTowardPositive | RoundTowardNegative | RoundTowardZero | Float of int * int | Fp of int * int | Plus_infinity of int * int | Minus_infinity of int * int | Plus_zero of int * int | Minus_zero of int * int | NaN of int * int | Fp_abs of int * int | Fp_neg of int * int | Fp_add of int * int | Fp_sub of int * int | Fp_mul of int * int | Fp_div of int * int | Fp_fma of int * int | Fp_sqrt of int * int | Fp_rem of int * int | Fp_roundToIntegral of int * int | Fp_min of int * int | Fp_max of int * int | Fp_leq of int * int | Fp_lt of int * int | Fp_geq of int * int | Fp_gt of int * int | Fp_eq of int * int | Fp_isNormal of int * int | Fp_isSubnormal of int * int | Fp_isZero of int * int | Fp_isInfinite of int * int | Fp_isNaN of int * int | Fp_isNegative of int * int | Fp_isPositive of int * int | Ieee_format_to_fp of int * int | Fp_to_fp of int * int * int * int | Real_to_fp of int * int | Sbv_to_fp of int * int * int | Ubv_to_fp of int * int * int | To_ubv of int * int * int | To_sbv of int * int * int | To_real of int * int

type builtin += String | Str of string | Str_length | Str_at | Str_to_code | Str_of_code | Str_is_digit | Str_to_int | Str_of_int | Str_concat | Str_sub | Str_index_of | Str_replace | Str_replace_all | Str_replace_re | Str_replace_re_all | Str_is_prefix | Str_is_suffix | Str_contains | Str_lexicographic_strict | Str_lexicographic_large | Str_in_re

type builtin += String_RegLan | Re_empty | Re_all | Re_allchar | Re_of_string | Re_range | Re_concat | Re_union | Re_inter | Re_star | Re_cross | Re_complement | Re_diff | Re_option | Re_power of int | Re_loop of int * int

Native Tags

module Tags : sig ... end

Filters

module Filter : sig ... end

Printing

module Print : sig ... end

Substitutions

module Subst : sig ... end

Identifiers

module Id : sig ... end

Types

module Ty : sig ... end

Terms

module Term : sig ... end