Module Term.Const

A module for constant symbols that occur in terms.

type t = term_const

The type of constant symbols that can occur in terms

val hash : t -> int

A hash function for term constants, should be suitable to create hashtables.

val equal : t -> t -> bool

An equality function on term constants. Should be compatible with the hash function.

val compare : t -> t -> int

Comparison function on variables.

val arity : t -> int * int

Returns the arity of a term constant.

val mk : string -> ty_var list -> ty list -> ty -> t

Create a polymorphic constant symbol.

val tag : t -> 'a tag -> 'a -> unit

Tag a constant.

val get_tag : t -> 'a tag -> 'a list

Return the list of values associated to the tag.

val get_tag_last : t -> 'a tag -> 'a option

Return the last value associated to the tag (if any).