Module Dolmen_std__Expr.Print

type 'a t = Stdlib.Format.formatter -> 'a -> unit

Alias for the type printing functions.

val print_index : bool Stdlib.ref

Determines whether to print the unique index of each identifier or not.

val name : Dolmen_std.Pretty.name Dolmen_std.Tag.t

The name tag is used for the printing of identifiers. When an identifier has an name tag, its value is used instead of the identifier intrinsic name.

val pos : Dolmen_std.Pretty.pos Dolmen_std.Tag.t

Positioning for pretty printing. If this tag is set, the printing functions will ignore type arguments (for readability). Pretty.Infix uses the identifier as a separator when printing th argument list Pretty.Prefix just ignore type arguments.

val id : _ id t

Printer for ids

val type_ : type_ t

Printer for type_.

val type_fun : type_fun t

Printer for type_fun.

val ty_var : ty_var t

Printer to print type variables along with their types.

val ty_cst : ty_cst t

Printer to print type constants along with their types.

val term_var : term_var t

Printer to print term variables along with their types.

val term_cst : term_cst t

Printer to print term constants along with their types.

val ty : ty t

Printer for types.

val term : term t

Printer for terms.

val formula : formula t

Printer for formulas.