Module Dolmen_std__Expr.Tags

type 'a t = 'a tag

Polymorphic tags

val bound : term tag

Tag used one let-bound variables to reference the defining term for the variable (i.e. the term to which it is let-bound).

Satsify the Smtlib interface.

include Dolmen_intf.Tag.Smtlib_Base with type 'a t := 'a t and type term := term

type term

The type of terms

type 'a t

Polymorphic tags

val named : string t

A tag used to named formulas in smtlib. Should correspond to the `:named` attribute.

val triggers : term list t

Multi-triggers (typically annotated on the body of a quantified formula and not the quantified formula itself).

val rwrt : unit t

A flag (i.e. unit tag), indicating that the tagged term/formula is to be considered as a rewrite rule.

Satsify the Zf interface.

include Dolmen_intf.Tag.Zf_Base with type 'a t := 'a t

type 'a t

Polymorphic tags

val rwrt : unit t

A flag (i.e. unit tag), indicatgin that the tagged term/formula is to be considered as a rewrite rule.

type name

A type used to specify a name for printing identifiers

val name : name t

A tag used to specify what to print when printing an identifier

val exact : string -> name

Print the identifier with this exact string.

type pos

A type to indicate how to print identifiers

val pos : pos t

A tag to specify how to print identifiers

val infix : pos

The tagged identifier is an infix symbol

val prefix : pos

The tagged identifier is a prefix symbol