Module Term.Var

A module for variables that occur in terms.

type t = term_var

The type of variables the can occur in terms

val print : Stdlib.Format.formatter -> t -> unit

Printer.

val hash : t -> int

A hash function for term variables, should be suitable to create hashtables.

val equal : t -> t -> bool

An equality function on term variables. Should be compatible with the hash function.

val compare : t -> t -> int

Comparison function on variables.

val mk : string -> ty -> t

Create a new typed variable.

val ty : t -> ty

Return the type of the variable.

val get_tag : t -> 'a Dolmen_std.Tag.t -> 'a option

Get the value bound to a tag.

val get_tag_list : t -> 'a list Dolmen_std.Tag.t -> 'a list

Get the list of values bound to a list tag, returning the empty list if no value is bound.

val get_tag_last : t -> 'a list Dolmen_std.Tag.t -> 'a option

Get the last value bound to a list tag.

val set_tag : t -> 'a Dolmen_std.Tag.t -> 'a -> unit

Set the value bound to the tag.

val add_tag : t -> 'a list Dolmen_std.Tag.t -> 'a -> unit

Bind an additional value to a list tag.

val add_tag_opt : t -> 'a list Dolmen_std.Tag.t -> 'a option -> unit

Optionally bind an additional value to a list tag.

val add_tag_list : t -> 'a list Dolmen_std.Tag.t -> 'a list -> unit

Bind a list of additional values to a list tag.

val unset_tag : t -> _ Dolmen_std.Tag.t -> unit

Remove the binding to the given tag.