Module Ty.Const

A module for constant symbols the occur in types.

type t = ty_const

The type of constant symbols the can occur in types

val hash : t -> int

A hash function for type constants, should be suitable to create hashtables.

val equal : t -> t -> bool

An equality function on type constants. Should be compatible with the hash function.

val compare : t -> t -> int

Comparison function on variables.

val arity : t -> int

Return the arity of the given symbol.

val mk : string -> int -> t

Create a type constant with the given arity.

val tag : t -> 'a tag -> 'a -> unit

Tag a variable.

val get_tag : t -> 'a tag -> 'a list

Return the list of values associated to the tag.

val get_tag_last : t -> 'a tag -> 'a option

Return the last value associated to the tag (if any).

val int : t

The type constant for integers

val rat : t

The type constant for rationals

val real : t

The type constant for reals.

val prop : t

The type constant for propositions

val unit : t

The unit type.

val base : t

An arbitrary type constant.

val array : t

The type constant for arrays

val bitv : int -> t

Bitvectors of the given length.

val string : t

The type constant for strings

val string_reg_lang : t

The type constant for regular languages over strings.