Module Dolmen_std__Term

Standard implementation of terms

Type definitions

type location = Dolmen_std.Loc.t

type builtin =

| Wildcard
Wildcard symbol, i.e placeholder for an expression to be inferred, typically during type-checking.

| Ttype
Builtin symbol for the type of Types.

| Unit
Unit type.

| Void
Only inhabitant of the unit type.

| Prop
Builtin symbol for the type of propositions.

| Bool
The boolean type.

| True
The true propositional constant.

| False
The false propositional constant.

| Eq
Should all arguments be pariwise equal ?

| Distinct
Should all arguments be pairwise distinct ?

| Ite
Condional, usually applied to 3 terms (the condition, the then branch and the else branch).

| Sequent
Sequent as term, usually takes two argument (left side, and right side of the sequent), which are respectively a conjunction and a disjunction of propositional formulas.

| Int
Builtin integer type. Currently specific to Zipperposition and alt-ergo format; other languages might use constants with pre-defined name, such as tptp's "$int".

| Real
Builtin real type. Currently specific to Alt-ergo format; other languages might use constants with pre-defined name, such as smtlib's "real".

| Minus
Arithmetic unary minus.

| Add
Arithmetic addition.

| Sub
Arithmetic substraction.

| Mult
Arithmetic multiplication.

| Div
Arithmetic division quotient

| Mod
Arithmetic division modulo

| Int_pow
Integer exponentiation.

| Real_pow
Real exponentiation.

| Lt
Arithmetic "less than" comparison (strict).

| Leq
Arithmetic "lesser or equal" comparison.

| Gt
Arithmetic "greater than" comparison.

| Geq
Arithmetic "greater or equal" comparison.

| Subtype
Subtyping relation

| Product
Product type constructor

| Union
Union type constructor

| Pi
Pi: higher-order encoding of the forall quantifier as a constant.

| Sigma
Sigma: higher-order envoding of the exists quantifier of a constant.

| Not
Propositional negation

| And
Propositional conjunction

| Or
Propositional disjunction

| Nand
Propositional not-and connective

| Xor
Propositional exclusive disjunction

| Nor
Propositional not-or

| Imply
Propositional implication

| Implied
Propositional left implication (i.e implication with reversed arguments).

| Equiv
Propositional equivalence

| Bitv of int
Bitvector type (with given length)

| Bitv_extract of int * int
Bitvector extraction

| Bitv_concat
Bitvector concatenation

| Array_get
Array getter.

| Array_set
Array setter.

| Adt_check
Algebraic datatype head constructore checker.

| Adt_project
Algebraic datatype projection.

| Record
Record creation

| Record_with
Record "with" creation

| Record_access
Record field access

| Maps_to
Mapping; used in Alt-ergo triggers.

| In_interval of bool * bool
Interval check; used in Alt-ergo triggers

| Check
Similar to cut, but does not introduce the proved term into the axioms.

| Cut
Insert a cut of the given term.

The type of builtins symbols for terms. Some languages have specific syntax for logical connectives (tptp's'&&' or '||' for isntance) whereas some (smtlib for instance) don't and treat them as constants.

type binder =

| All
Universal quantification. Each term in the list of quantified terms should represent a variable (optionnally typed using the Colon constructor.

| Ex
Existencial quantification Each term in the list of quantified terms should represent a variable (optionnally typed using the Colon constructor.

| Pi
Polymorphic type quantification in function type Each term in the list of quantified terms should represent a variable (optionnally typed using the Colon constructor.

| Arrow
The arrow binder, for function types. Allows for curified types, if wanted.

| Let_seq
Let bindings (either propositional or for terms). Term bound by a let can have many forms depending on the language, but usual shapes are:

an equality (using the builtin Eq) between a variable (optionnally typed using the Colon constructor), and a term (e.g. in tptp)

an equivalence (using the builtin Equiv) between a variable (optionnally typed using the Colon constructor), and a term/proposition (e.g. in tptp)

a variable and a term juxtaposed using the Colon constructor (e.g. in smtlib)

| Let_par
Similar to Let_seq; except that the list of bindings should be considered all bound at the same time/level/scope. More precisely, for Let_seq, the list of bindings is to be understood sequentially (i.e. Let_seq (b1 :: b2 ...) is semantically the same as Let_seq b1 (Let_seq b2 (..)). For Let_par, the list of bindings all happen at the same time: the defining expressions of each binding cannot refer to other bindings in the same parralel let-binding.

| Fun
Lambda, i.e function abstraction binder. Bound terms are the variables bound by the lambda, optionnally typed using the Colon constructor.

| Choice
Indefinite description, or epsilon terms. Likely to have its usual shape change following tptp's recent changes.

| Description
Definite description. Likely to have its usual shape change following tptp's recent changes.

The type of binders, these are pretty much always builtin in all languages.

type descr =

| Symbol of Dolmen_std.Id.t
Constants, variables, etc... any string-identified non-builtin atomic term.

| Builtin of builtin
Predefined builtins, i.e constants with lexical or syntaxic defintion in the source language.

| Colon of t * t
Juxtaposition of terms, usually used to annotate a term with its type (for quantified variables, functions arguments, etc...).

| App of t * t list
Higher-order application

| Binder of binder * t list * t
Binder (quantifiers, local functions, ...), see the binder type for more information.

| Match of t * (t * t) list
Pattern matching, the list contains tuples of the form (pattern,branch).

The AST for terms

and t = {

term : descr;

attr : t list;

loc : location;

}

The type of terms. A record containing an optional location, and a description of the term.

Standard functions

val equal : t -> t -> bool

val compare : t -> t -> int

Equality and comparison

val print : Stdlib.Format.formatter -> t -> unit

val print_builtin : Stdlib.Format.formatter -> builtin -> unit

Printing functionson buffer and formatters.

Implemented interfaces

Include the Logic interface. This interface defines almost all term building functions that you may want to use.

include Dolmen_intf.Term.Logic with type t := t and type id := Dolmen_std.Id.t and type location := location

type t

The type of terms.

type id

The type of identifiers used for constants.

type location

The type of locations attached to terms.

Predefined terms

val eq_t : ?⁠loc:location -> unit -> t

val neq_t : ?⁠loc:location -> unit -> t

The terms representing equality and disequality, respectively.

val wildcard : ?⁠loc:location -> unit -> t

The wildcard term, usually used in place of type arguments to explicit polymorphic functions to not explicit types that can be inferred by the type-checker.

val tType : ?⁠loc:location -> unit -> t

The type of types, defined as specific token by the Zipperposition format; in other languages, will be represented as a constant (the "$tType" constant in tptp for instance). Used to define new types, or quantify type variables in languages that support polymorphism.

val prop : ?⁠loc:location -> unit -> t

The type of propositions. Also defined as a lexical token by the Zipperposition format. Will be defined as a constant in most other languages (for instance, "$o" in tptp).

val bool : ?⁠loc:location -> unit -> t

The type of boolean, defined as a specific token by the Alt-ergo format; in other languages, it might be represented as a constant with a specific name.

val ty_unit : ?⁠loc:location -> unit -> t

The type unit, defined as a specific token by the Alt-ergo format; in other languages, it might be represented as a constant with a specific name.

val ty_int : ?⁠loc:location -> unit -> t

The type of integers, defined as a specific token by the Zipperposition and Alt-ergo formats; in other languages, it might be represented as a constant with a specific name (for isntance, tptp's "$int") .

val ty_real : ?⁠loc:location -> unit -> t

The type of integers, defined as a specific token by the Alt-ergo format; in other languages, it might be represented as a constant with a specific name (for isntance, tptp's "$int") .

val ty_bitv : ?⁠loc:location -> int -> t

The type of bitvectors of the given constant length, defined as a specifi token by the Alt-ergo format; in other languages, it might be represented as a constant with a specific name (for isntance, smtlib(s "bitv") .

val void : ?⁠loc:location -> unit -> t

The only value of type unit, defined as a specific token by the Alt-ergo format.

val true_ : ?⁠loc:location -> unit -> t

val false_ : ?⁠loc:location -> unit -> t

The constants for the true and false propositional constants. Again defined as lexical token in the Zipperposition format, while treated as a constant in other languages ("$true" in tptp).

val not_t : ?⁠loc:location -> unit -> t

val or_t : ?⁠loc:location -> unit -> t

val and_t : ?⁠loc:location -> unit -> t

val xor_t : ?⁠loc:location -> unit -> t

val nor_t : ?⁠loc:location -> unit -> t

val nand_t : ?⁠loc:location -> unit -> t

val equiv_t : ?⁠loc:location -> unit -> t

val implied_t : ?⁠loc:location -> unit -> t

val implies_t : ?⁠loc:location -> unit -> t

val pi_t : ?⁠loc:location -> unit -> t

val sigma_t : ?⁠loc:location -> unit -> t

Standard logical connectives viewed as terms. implies_t is usual right implication, i.e apply implies_t [p; q] is "p implies q", while apply implied_t [p; q ] means "p is implied by q" or "q implies p".

val data_t : ?⁠loc:location -> unit -> t

Term without semantic meaning, used for creating "data" terms. Used in tptp's annotations, and with similar meaning as smtlib's s-expressions (as used in the sexpr function defined later).

Terms leaf constructors

val var : ?⁠loc:location -> id -> t

val const : ?⁠loc:location -> id -> t

Variable and constant constructors. While in some languages they can distinguished at the lexical level (in tptp for instance), in most languages, it is an issue dependant on scoping rules, so terms parsed from an smtlib file will have all variables parsed as constants.

val atom : ?⁠loc:location -> int -> t

Atoms are used for dimacs cnf parsing. Positive integers denotes variables, and negative integers denote the negation of the variable corresponding to their absolute value.

val distinct : ?⁠loc:location -> id -> t

Used in tptp to specify constants different from other constants, for instance the 'distinct' "Apple" should be syntactically different from the "Apple" constant. Can be safely aliased to the const function as the distinct function is always given strings already enclosed with quotes, so in the example above, const would be called with "Apple" as string argument, while distinct would be called with the string "\"Apple\""

val str : ?⁠loc:location -> string -> t

val int : ?⁠loc:location -> string -> t

val rat : ?⁠loc:location -> string -> t

val real : ?⁠loc:location -> string -> t

val hexa : ?⁠loc:location -> string -> t

val binary : ?⁠loc:location -> string -> t

Constructors for words defined as numeric or string formats by the languages specifications. These also can be safely aliased to const, but then the provenance information is lost, which might complicate the task of a type-checker.

val bitv : ?⁠loc:location -> string -> t

Bitvetor litteral, defined as a specific token in Alt-ergo; Expects a decimal integer in the string to be extended as a bitvector.

Term constructors

val colon : ?⁠loc:location -> t -> t -> t

Represents juxtaposition of two terms, usually denoted "t : t'" in most languages, and mainly used to annotated terms with their supposed, or defined, type.

val eq : ?⁠loc:location -> t -> t -> t

val neq : ?⁠loc:location -> t list -> t

Equality and dis-equality of terms.

val not_ : ?⁠loc:location -> t -> t

val or_ : ?⁠loc:location -> t list -> t

val and_ : ?⁠loc:location -> t list -> t

val xor : ?⁠loc:location -> t -> t -> t

val imply : ?⁠loc:location -> t -> t -> t

val equiv : ?⁠loc:location -> t -> t -> t

Proposition construction functions. The conjunction and disjunction are n-ary instead of binary mostly because they are in smtlib (and that is subsumes the binary case).

val apply : ?⁠loc:location -> t -> t list -> t

Application constructor, seen as higher order application rather than first-order application for the following reasons: being able to parse tptp's THF, having location attached to function symbols.

val ite : ?⁠loc:location -> t -> t -> t -> t

Conditional constructor, both for first-order terms and propositions. Used in the following schema: ite condition then_branch else_branch.

val match_ : ?⁠loc:location -> t -> (t * t) list -> t

Pattern matching. The first term is the term to match, and each tuple in the list is a match case, which is a pair of a pattern and a match branch.

val pi : ?⁠loc:location -> t list -> t -> t

val par : ?⁠loc:location -> t list -> t -> t

val letin : ?⁠loc:location -> t list -> t -> t

val letand : ?⁠loc:location -> t list -> t -> t

val forall : ?⁠loc:location -> t list -> t -> t

val exists : ?⁠loc:location -> t list -> t -> t

val lambda : ?⁠loc:location -> t list -> t -> t

val choice : ?⁠loc:location -> t list -> t -> t

val description : ?⁠loc:location -> t list -> t -> t

Binders for variables. Takes a list of terms as first argument for simplicity, the lists will almost always be a list of variables, optionally typed using the colon term constructor.

Pi is the polymorphic type quantification, for instance the polymorphic identity function has type: "Pi alpha. alpha -> alpha"

Letin is local binding, takes a list of equality of equivalences whose left hand-side is a variable. Letand is the parrallel version of Letin.

Forall is universal quantification

Par is universal quantification over type variables specifically (i.e. the same as forall, but only for a list of type variables, which thus may omit the colon annotations in the arguments).

Exists is existential quantification

Lambda is used for function construction

Choice is the choice operator, also called indefinite description, or also epsilon terms, i.e "Choice x. p(x)" is one "x" such that "p(x)" is true.

Description is the definite description, i.e "Description x. p(x)" is the only "x" that satisfies p.

Type constructors

val arrow : ?⁠loc:location -> t -> t -> t

Function type constructor, for curryfied functions. Functions that takes multiple arguments in first-order terms might take a product as only argument (see the following product function) in some languages (e.g. tptp), or be curryfied using this constructor in other languages (e.g. alt-ergo).

val product : ?⁠loc:location -> t -> t -> t

Product type constructor, used for instance in the types of functions that takes multiple arguments in a non-curry way.

val union : ?⁠loc:location -> t -> t -> t

Union type constructor, currently used in tptp's THF format.

val subtype : ?⁠loc:location -> t -> t -> t

Subtype relation for types.

Record constructors

val record : ?⁠loc:location -> t list -> t

Create a record expression.

val record_with : ?⁠loc:location -> t -> t list -> t

Record "with" update (e.g. " r with ....").

val record_access : ?⁠loc:location -> t -> id -> t

Field record access.

Algebraic datatypes

val adt_check : ?⁠loc:location -> t -> id -> t

Check whether some expression matches a given adt constructor (in head position).

val adt_project : ?⁠loc:location -> t -> id -> t

Project a field of an adt constructor (usually unsafe except when guarded by an adt_check function).

Array constructors

val array_get : ?⁠loc:location -> t -> t -> t

Array getter.

val array_set : ?⁠loc:location -> t -> t -> t -> t

Array setter.

Bitvector constructors

val bitv_extract : ?⁠loc:location -> t -> int -> int -> t

Bitvector extraction.

val bitv_concat : ?⁠loc:location -> t -> t -> t

Bitvector concatenation.

Arithmetic constructors

val uminus : ?⁠loc:location -> t -> t

Arithmetic unary minus.

val add : ?⁠loc:location -> t -> t -> t

Arithmetic addition.

val sub : ?⁠loc:location -> t -> t -> t

Arithmetic substraction.

val mult : ?⁠loc:location -> t -> t -> t

Arithmetic multiplication.

val div : ?⁠loc:location -> t -> t -> t

Arithmetic division quotient.

val mod_ : ?⁠loc:location -> t -> t -> t

Arithmetic modulo (aka division reminder).

val int_pow : ?⁠loc:location -> t -> t -> t

Integer power.

val real_pow : ?⁠loc:location -> t -> t -> t

Real power.

val lt : ?⁠loc:location -> t -> t -> t

Arithmetic "lesser than" comparison (strict).

val leq : ?⁠loc:location -> t -> t -> t

Arithmetic "lesser or equal" comparison.

val gt : ?⁠loc:location -> t -> t -> t

Arithmetic "greater than" comparison (strict).

val geq : ?⁠loc:location -> t -> t -> t

Arithmetic "greater or equal" comparison.

Triggers

val in_interval : ?⁠loc:location -> t -> (t * bool) -> (t * bool) -> t

Create a predicate for whether a term is within the given bounds (each bound is represented by a term which is tis value and a boolean which specifies whether it is strict or not).

val maps_to : ?⁠loc:location -> id -> t -> t

Id mapping (see alt-ergo).

val trigger : ?⁠loc:location -> t list -> t

Create a multi-trigger (i.e. all terms in the lsit must match to trigger).

val triggers : ?⁠loc:location -> t -> t list -> t

triggers ~loc f l annotates formula/term f with a list of triggers.

val filters : ?⁠loc:location -> t -> t list -> t

filters ~loc f l annotates formula/term f with a list of filters.

Special constructions

val tracked : ?⁠loc:location -> id -> t -> t

Name a term for tracking purposes.

val quoted : ?⁠loc:location -> string -> t

Create an attribute from a quoted string (in Zf).

val sequent : ?⁠loc:location -> t list -> t list -> t

Sequents as terms

val check : ?⁠loc:location -> t -> t

Check a term (see alt-ergo).

val cut : ?⁠loc:location -> t -> t

Create a cut (see alt-ergo).

val annot : ?⁠loc:location -> t -> t list -> t

Attach a list of attributes (also called annotations) to a term. Attributes have no logical meaning (they can be safely ignored), but may serve to give hints or meta-information.

val sexpr : ?⁠loc:location -> t list -> t

S-expressions (for smtlib attributes), should probably be related to the data_t term.

Term constructor not in implemented interfaces

val ite_t : ?⁠loc:location -> unit -> t

The standalone term corresponding to the if-then-else builtin construction.

Term inspection

val fv : t -> Dolmen_std.Id.t list

Return the list of free variables (i.e currently, Ids that are in the Var namespace).

Additional functions

val builtin : builtin -> ?⁠loc:location -> unit -> t

Make a builtin.

val fun_ty : ?⁠loc:location -> t list -> t -> t

Multi-arguments function type constructor.

val add_attr : t -> t -> t

add_attr attr term rturns a term t equal to term, but with attr added to the list of attributes.

val add_attrs : t list -> t -> t

Same as add_attr but adds a list of attributes.

val set_attrs : t list -> t -> t

Set the given list of terms as th attributes of the given term. Will fail (with an assertion) if the given term already have some assertion. In such cases, use add_attr instead.

Term mapping

The main use of terms mapper is to map fuctions over some terms. Traditionally, a mapping will usually only care about a few syntax cases and leav all other untouched. In these cases, it is useful to override the identity mapper, redefining only the fields needed.

type 'a mapper = {

symbol : 'a mapper -> attr:t list -> loc:location -> Dolmen_std.Id.t -> 'a;

builtin : 'a mapper -> attr:t list -> loc:location -> builtin -> 'a;

colon : 'a mapper -> attr:t list -> loc:location -> t -> t -> 'a;

app : 'a mapper -> attr:t list -> loc:location -> t -> t list -> 'a;

binder : 'a mapper -> attr:t list -> loc:location -> binder -> t list -> t -> 'a;

pmatch : 'a mapper -> attr:t list -> loc:location -> t -> (t * t) list -> 'a;

}

The type of a mapper on terms.

val id_mapper : t mapper

The identity mapper: maps any term to itself.

val unit_mapper : unit mapper

The unit mapper, i.e. an iterator.

val map : 'a mapper -> t -> 'a

Apply a mapper to a term.